refactor: move function selector methods out of `ABIParser` struct
* This methods will ultimately be used inside the Sway repository, and
  they didn't use the ABIParser fields.
