+++ packages/fuels-core/src/code_gen/functions_gen.rs
@@ -1,10 +1,10 @@
 use crate::code_gen::custom_types_gen::extract_custom_type_name_from_abi_property;
 use crate::code_gen::docs_gen::expand_doc;
-use crate::json_abi::ABIParser;
 use crate::types::expand_type;
 use crate::utils::{first_four_bytes_of_sha256_hash, ident, safe_ident};
 use crate::{ParamType, Selector};
 use fuels_types::errors::Error;
+use fuels_types::function_selector::build_fn_selector;
 use fuels_types::{CustomType, Function, Property, ENUM_KEYWORD, STRUCT_KEYWORD};
 use inflector::Inflector;
 use proc_macro2::{Literal, TokenStream};
@@ -24,7 +24,6 @@ use std::collections::HashMap;
 /// [`Contract`]: crate::contract::Contract
 pub fn expand_function(
     function: &Function,
-    abi_parser: &ABIParser,
     custom_enums: &HashMap<String, Property>,
     custom_structs: &HashMap<String, Property>,
 ) -> Result<TokenStream, Error> {
@@ -33,7 +32,7 @@ pub fn expand_function(
     }
 
     let name = safe_ident(&function.name);
-    let fn_signature = abi_parser.build_fn_selector(&function.name, &function.inputs)?;
+    let fn_signature = build_fn_selector(&function.name, &function.inputs)?;
 
     let encoded = first_four_bytes_of_sha256_hash(&fn_signature);
 
@@ -372,12 +371,7 @@ mod tests {
             type_field: String::from("bool"),
             components: None,
         });
-        let result = expand_function(
-            &the_function,
-            &ABIParser::new(),
-            &Default::default(),
-            &Default::default(),
-        );
+        let result = expand_function(&the_function, &Default::default(), &Default::default());
         let expected = TokenStream::from_str(
             r#"
             #[doc = "Calls the contract's `HelloWorld` (0x0000000097d4de45) function"]
@@ -465,8 +459,7 @@ mod tests {
                 components: None,
             },
         );
-        let abi_parser = ABIParser::new();
-        let result = expand_function(&the_function, &abi_parser, &custom_enums, &custom_structs);
+        let result = expand_function(&the_function, &custom_enums, &custom_structs);
         // Some more editing was required because it is not rustfmt-compatible (adding/removing parentheses or commas)
         let expected = TokenStream::from_str(
             r#"
