+++ packages/fuels-types/src/lib.rs
@@ -8,8 +8,10 @@ use strum_macros::ToString;
 
 pub mod constants;
 pub mod errors;
+pub mod function_selector;
 pub mod param_types;
 pub mod parse_param;
+pub mod utils;
 /// Fuel ABI representation in JSON, originally specified here:
 ///
 /// https://github.com/FuelLabs/fuel-specs/blob/master/specs/protocol/abi.md
